services/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn>=0.43.0",
    "pydantic-settings>=2.7.0",
    "asyncpg>=0.30.0",
    "redis[hiredis]>=5.2.0",
    "shared",
]

[tool.uv.sources]
shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
